Move calculation of slice item in NDCube.crop to utils.
This enables the item to be calculated without slicing the cube which
can be useful, for example when trying to mimic crop's behaviour for
NDCubeSequence.
