+++ ndcube/utils/misc.py
@@ -1,11 +1,6 @@
-import inspect
-from functools import wraps
-
 import astropy.units as u
-import numpy as np
-from astropy.wcs.wcsapi import BaseHighLevelWCS
 
-__all__ = ['sanitise_wcs', 'unique_sorted']
+__all__ = ['unique_sorted']
 
 
 def unique_sorted(iterable):
@@ -14,86 +9,6 @@ def unique_sorted(iterable):
     """
     lookup = set()  # a temporary lookup set
     return [ele for ele in iterable if ele not in lookup and lookup.add(ele) is None]
-
-
-def sanitise_wcs(func):
-    """
-    A wrapper for NDCube methods to sanitise the wcs argument.
-
-    This decorator is only designed to be used on methods of NDCube.
-
-    It will find the wcs argument, keyword or positional and if it is None, set
-    it to `self.wcs`.
-    It will then verify that the WCS has a matching number of pixel dimensions
-    to the dimensionality of the array. It will finally verify that the object
-    passed is a HighLevelWCS object, or an ExtraCoords object.
-    """
-    # This needs to be here to prevent a circular import
-    from ndcube.extra_coords import ExtraCoords
-
-    @wraps(func)
-    def wcs_wrapper(*args, **kwargs):
-        sig = inspect.signature(func)
-        params = sig.bind(*args, **kwargs)
-        wcs = params.arguments.get('wcs', None)
-        self = params.arguments['self']
-
-        if wcs is None:
-            wcs = self.wcs
-
-        if not isinstance(wcs, ExtraCoords):
-            if not wcs.pixel_n_dim == self.data.ndim:
-                raise ValueError("The supplied WCS must have the same number of "
-                                 "pixel dimensions as the NDCube object. "
-                                 "If you specified `cube.extra_coords.wcs` "
-                                 "please just pass `cube.extra_coords`.")
-
-        if not isinstance(wcs, (BaseHighLevelWCS, ExtraCoords)):
-            raise TypeError("wcs argument must be a High Level WCS or an ExtraCoords object.")
-
-        params.arguments['wcs'] = wcs
-
-        return func(*params.args, **params.kwargs)
-
-    return wcs_wrapper
-
-
-def sanitize_crop_inputs(lower_corner, upper_corner, wcs):
-    """Sanitize inputs to NDCube crop methods.
-
-    First arg returned signifies whether the inputs imply that cropping
-    should be performed or not.
-    """
-    lower_corner, upper_corner = sanitize_corners(lower_corner, upper_corner)
-
-    # Quit out early if we are no-op
-    lower_nones = np.array([lower is None for lower in lower_corner])
-    upper_nones = np.array([upper is None for upper in upper_corner])
-    if (lower_nones & upper_nones).all():
-        return True, lower_corner, upper_corner, wcs
-
-    # This needs to be here to prevent a circular import
-    from ndcube.extra_coords import ExtraCoords
-    if isinstance(wcs, ExtraCoords):
-        # Add None inputs to upper and lower corners for new dummy axes.
-        n_dummy_axes = len(wcs._cube_array_axes_without_extra_coords)
-        lower_corner += [None] * n_dummy_axes
-        upper_corner += [None] * n_dummy_axes
-        # Convert extra coords to WCS describing whole cube.
-        wcs = wcs.cube_wcs
-
-    return False, lower_corner, upper_corner, wcs
-
-
-def sanitize_corners(*corners):
-    """Sanitize corner inputs to NDCube crop methods."""
-    corners = [list(corner) if isinstance(corner, (tuple, list)) else [corner]
-               for corner in corners]
-    n_coords = [len(corner) for corner in corners]
-    if len(set(n_coords)) != 1:
-        raise ValueError("All corner inputs must have same number of coordinate objects. "
-                         f"Lengths of corner objects: {n_coords}")
-    return corners
 
 
 def convert_quantities_to_units(coords, units):
